v2 api for store admin authentication
