--- street-goods-burger-backend/app/controllers/store_admin/store_admin_controller.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-# StoreAdmin::StoreAdminController Template
-class StoreAdmin::StoreAdminController < ApplicationController
-  before_action :authenticate_store_admin!
-end
