--- street-goods-burger-backend/app/controllers/store_admin/view_registered_store_customers_controller.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-# StoreAdmin::ViewRegisteredStoreCustomerController Template
-class StoreAdmin::ViewRegisteredStoreCustomersController < StoreAdmin::StoreAdminController
-  def index
-    store_id = current_store_admin[:id]
-    view_registered_store_customer_info = current_store_admin
-                                          .view_registered_store_customers(store_id:)
-    render json: view_registered_store_customer_info
-  end
-end
