--- street-goods-burger-backend/app/controllers/store_customer/create_store_transaction_controller.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-# StoreCustomer::CreateStoreTransaction Template
-class StoreCustomer::CreateStoreTransactionController < StoreCustomer::StoreCustomerController
-  def create
-    create_store_transaction_info = current_store_customer
-                                    .create_store_transaction
-    render json: create_store_transaction_info
-  end
-end
