--- street-goods-burger-backend/app/controllers/store_customer/like_food_controller.rb
@@ -1,25 +1,0 @@
-# frozen_string_literal: true
-
-# StoreCustomer::LikeFoodController Template
-class StoreCustomer::LikeFoodController < StoreCustomer::StoreCustomerController
-  def update
-    food_id = like_food_params[:food_id]
-    like_food_info = current_store_customer.like_food(food_id:)
-
-    render json: like_food_info
-  rescue ActionController::ParameterMissing => e
-    render json: {
-      status: 400,
-      message: 'invalid parameters',
-      errors: e.message
-    }
-  end
-
-  private
-
-  def like_food_params
-    params
-      .require(:food_info)
-      .permit(:food_id)
-  end
-end
