--- street-goods-burger-backend/app/controllers/store_customer/view_favorite_foods_controller.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-# StoreCustomer::ViewFavoriteFoodController Template
-class StoreCustomer::ViewFavoriteFoodsController < StoreCustomer::StoreCustomerController
-  def index
-    view_favorite_food_info = current_store_customer.view_favorite_foods
-    render json: view_favorite_food_info
-  end
-end
